\name{FastHCS-package}
\docType{package}
\title{Code to compute the FastHCS outlyingness index for high-dimensional data-sets.}
\description{
Uses the FastHCS algorithm to compute the HCS outlyingness index 
 of regression.
}
\details{
\tabular{ll}{
Package: \tab FastHCS\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-01-13\cr
Suggests: \tab mvtnorm\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
FastHCS                 Function to compute the FastHCS outlyingness index for 
			high-dimensional data-sets.
computdis                 Internal function used to compute the FastHCS outlyingness index for 
			high-dimensional data-sets.
NumStarts                 Internal function used to compute the FastHCS outlyingness index for 
			high-dimensional data-sets.
quanf                 Internal function used to compute the FastHCS outlyingness index for 
			high-dimensional data-sets.
Signflip                 Internal function used to compute the FastHCS outlyingness index for 
			high-dimensional data-sets.
}
}
\author{
Kaveh Vakili [aut, cre],
Maintainer: Kaveh Vakili <kaveh.vakili@wis.kuleuven.be>
}
\keyword{package}
